refactor(server): pages middleware
