+++ pages/profile/_middleware.ts
@@ -1,15 +1,1 @@
-import type { NextMiddleware } from "next/server";
-import { NextResponse } from "next/server";
-
-const ProfileMiddleware: NextMiddleware = (request) => {
-  const { auth } = request.cookies;
-
-  if (!auth) {
-    const url = request.nextUrl.clone();
-    url.pathname = "/login";
-
-    return NextResponse.redirect(url);
-  }
-};
-
-export { ProfileMiddleware as middleware };
+export { MustBeAuthedMiddleware as middleware } from "../../src/server/utils/middleware";
