+++ src/server/utils/middleware/index.ts
@@ -1,0 +1,1 @@
+export * from "./auth.middleware";
